+++ src/main/java/dz/djezzy/site/acceptance/business/repository/AuditSiteLineRepository.java
@@ -3,5 +3,5 @@ package dz.djezzy.site.acceptance.business.repository;
 import dz.djezzy.site.acceptance.business.data.entities.AuditSiteLine;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuditSiteLineRepository extends JpaRepository<AuditSiteLine, Long> {
+public interface AuditSiteLineRepository extends JpaRepository<AuditSiteLine, Integer> {
 }
