+++ src/main/java/dz/djezzy/site/acceptance/business/repository/AuditSiteRepository.java
@@ -3,5 +3,5 @@ package dz.djezzy.site.acceptance.business.repository;
 import dz.djezzy.site.acceptance.business.data.entities.AuditSite;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuditSiteRepository extends JpaRepository<AuditSite, Long> {
+public interface AuditSiteRepository extends JpaRepository<AuditSite, Integer> {
 }
