+++ src/main/java/dz/djezzy/site/acceptance/business/repository/CategoriesRepository.java
@@ -3,5 +3,5 @@ package dz.djezzy.site.acceptance.business.repository;
 import dz.djezzy.site.acceptance.business.data.entities.Categories;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoriesRepository extends JpaRepository<Categories, Long> {
+public interface CategoriesRepository extends JpaRepository<Categories, Integer> {
 }
