+++ src/main/java/dz/djezzy/site/acceptance/business/repository/SiteRepository.java
@@ -3,5 +3,5 @@ package dz.djezzy.site.acceptance.business.repository;
 import dz.djezzy.site.acceptance.business.data.entities.Site;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SiteRepository extends JpaRepository<Site, Long> {
+public interface SiteRepository extends JpaRepository<Site, Integer> {
 }
