+++ src/main/java/dz/djezzy/site/acceptance/business/repository/StatusAuditSiteRepository.java
@@ -3,5 +3,5 @@ package dz.djezzy.site.acceptance.business.repository;
 import dz.djezzy.site.acceptance.business.data.entities.StatusAuditSite;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StatusAuditSiteRepository extends JpaRepository<StatusAuditSite, Long> {
+public interface StatusAuditSiteRepository extends JpaRepository<StatusAuditSite, Integer> {
 }
