+++ src/main/java/dz/djezzy/site/acceptance/business/repository/StatusRepository.java
@@ -3,5 +3,5 @@ package dz.djezzy.site.acceptance.business.repository;
 import dz.djezzy.site.acceptance.business.data.entities.Status;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StatusRepository extends JpaRepository<Status, Long> {
+public interface StatusRepository extends JpaRepository<Status, Integer> {
 }
