+++ src/main/java/dz/djezzy/site/acceptance/business/services/AuditSiteLineService.java
@@ -3,5 +3,5 @@ package dz.djezzy.site.acceptance.business.services;
 import dz.djezzy.site.acceptance.business.data.dto.AuditSiteLineDto;
 import dz.djezzy.site.acceptance.business.data.entities.AuditSiteLine;
 
-public interface AuditSiteLineService extends GenericService<AuditSiteLine, AuditSiteLineDto, Long> {
+public interface AuditSiteLineService extends GenericService<AuditSiteLine, AuditSiteLineDto, Integer> {
 }
