+++ src/main/java/dz/djezzy/site/acceptance/business/services/AuditSiteService.java
@@ -3,5 +3,5 @@ package dz.djezzy.site.acceptance.business.services;
 import dz.djezzy.site.acceptance.business.data.dto.AuditSiteDto;
 import dz.djezzy.site.acceptance.business.data.entities.AuditSite;
 
-public interface AuditSiteService extends GenericService<AuditSite, AuditSiteDto, Long> {
+public interface AuditSiteService extends GenericService<AuditSite, AuditSiteDto, Integer> {
 }
