+++ src/main/java/dz/djezzy/site/acceptance/business/services/CategoriesService.java
@@ -3,5 +3,5 @@ package dz.djezzy.site.acceptance.business.services;
 import dz.djezzy.site.acceptance.business.data.dto.CategoriesDto;
 import dz.djezzy.site.acceptance.business.data.entities.Categories;
 
-public interface CategoriesService extends GenericService<Categories, CategoriesDto, Long> {
+public interface CategoriesService extends GenericService<Categories, CategoriesDto, Integer> {
 }
