+++ src/main/java/dz/djezzy/site/acceptance/business/services/SiteService.java
@@ -3,5 +3,5 @@ package dz.djezzy.site.acceptance.business.services;
 import dz.djezzy.site.acceptance.business.data.dto.SiteDto;
 import dz.djezzy.site.acceptance.business.data.entities.Site;
 
-public interface SiteService extends GenericService<Site, SiteDto, Long> {
+public interface SiteService extends GenericService<Site, SiteDto, Integer> {
 }
