+++ src/main/java/dz/djezzy/site/acceptance/business/services/StatusAuditSiteService.java
@@ -3,5 +3,5 @@ package dz.djezzy.site.acceptance.business.services;
 import dz.djezzy.site.acceptance.business.data.dto.StatusAuditSiteDto;
 import dz.djezzy.site.acceptance.business.data.entities.StatusAuditSite;
 
-public interface StatusAuditSiteService extends GenericService<StatusAuditSite, StatusAuditSiteDto, Long> {
+public interface StatusAuditSiteService extends GenericService<StatusAuditSite, StatusAuditSiteDto, Integer> {
 }
