+++ src/main/java/dz/djezzy/site/acceptance/business/services/StatusService.java
@@ -3,5 +3,5 @@ package dz.djezzy.site.acceptance.business.services;
 import dz.djezzy.site.acceptance.business.data.dto.StatusDto;
 import dz.djezzy.site.acceptance.business.data.entities.Status;
 
-public interface StatusService extends GenericService<Status, StatusDto, Long> {
+public interface StatusService extends GenericService<Status, StatusDto, Integer> {
 }
