+++ src/main/java/dz/djezzy/site/acceptance/business/services/SubCategoriesService.java
@@ -3,5 +3,5 @@ package dz.djezzy.site.acceptance.business.services;
 import dz.djezzy.site.acceptance.business.data.dto.SubCategoriesDto;
 import dz.djezzy.site.acceptance.business.data.entities.SubCategories;
 
-public interface SubCategoriesService extends GenericService<SubCategories, SubCategoriesDto, Long> {
+public interface SubCategoriesService extends GenericService<SubCategories, SubCategoriesDto, Integer> {
 }
